pyproject.toml
[tool.poetry]
name = "ai agent platform"
version = "0.1.0"
description = ""
authors = ["BakrFrag"]
#readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.121.3"}
sqlalchemy = "^2.0.44"
alembic = "^1.17.2"
aiosqlite = "^0.21.0"
openapi = "^2.0.0"
python-magic = "^0.4.27"
openai = {version = ">=1.0.0b1", allow-prereleases = true}
uuid-utils = "^0.11.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
